fix: rename component to example
